pyproject.toml
[project]
name = "minecraft-mod-agent-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "strands-agents[openai]>=1.26.0",
]
